design: mobile header stacking and quieter new-user Today
PageHeader stacks the aside above the kicker on small screens so the
countdown never squeezes the greeting; insights are suppressed until the
diagnostic exists so the hero is the only story for new users; short
date format in the greeting.

+++ src/app/(app)/page.tsx
@@ -22,8 +22,8 @@ export default async function DashboardPage() {
   const { summary, insight } = dash;
 
   const today = new Date().toLocaleDateString("en-US", {
-    weekday: "long",
-    month: "long",
+    weekday: "short",
+    month: "short",
     day: "numeric",
   });
   const first = user.name?.split(" ")[0];
